Create module for BookObject class, export and import
